[Leadership Reports] Refactor getBoundsDateString (Recidiviz/recidiviz-data#36654)
## Description of the change

Refactor `getBoundsDateString` to take in 1 string and return 1 string.
Additionally, change `previousMonthString` to `startDatePlusWeekString`
for the the MAU by Week query.

The `startDatePlusWeekString` only needs to be the start_date plus 1
week in order for the MAU by week query to work for both the previous
month (ex: December) and the previous month - 1 (ex: November).

### How I Tested
I produced the following
[November](https://docs.google.com/document/d/1N6McVC4AMujUV0aAdMgNUQy7yrwp6MLnoJHLC0oYOkk/edit?tab=t.0#heading=h.7tap3soi8gte)
and
[December](https://docs.google.com/document/d/1fEawGUvBTv8NxW67WVA61rloQ_THtdIh6Hi0uNsCF_E/edit?tab=t.0#heading=h.7tap3soi8gte)
reports. This fixes the December
[report](https://docs.google.com/document/d/1MdbZX0ZiyLteEAMRxOTr4lGSKO5_CtSAkpRSDCvd_iM/edit?tab=t.0)
as before the chart looked as follows:
![Screenshot 2024-12-23 at 2 05
41 PM](https://github.com/user-attachments/assets/e28f11b5-d4ad-4410-a600-5a1ea16162a9)

and now it looks like:
![Screenshot 2024-12-23 at 2 07
19 PM](https://github.com/user-attachments/assets/68847a55-b1dd-4d7e-a20e-c6f0754df1da)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

Co-authored-by: Michelle Orden <michelleorden@Michelles-MacBook-Pro.local>
GitOrigin-RevId: f462836ee22d617471d8f607990f74464f75769a