[ID Perf] add (#) months flagged as an outlier (Recidiviz/recidiviz-data#36629)
## Description of the change

This adds Yes/No for whether we flagged someone as an outlier in a given
month, and puts the number of "Yes"s in the all of 2024 column.

This change also required loading a sandbox, since in prod outliers we
only calculate 6 months worth of outliers data. Using the archives would
have been another option, but it gets a bit hairy since the archives
could say that in November we surfaced someone as an outlier in October,
but then if we look at production Insights, the historical rate chart
might not show them as having been an outlier in October (even though we
did technically surface them in the tool then). In order to maintain
consistency between the tool and these reports, I opted to go the
sandbox route. Since the aggregated metrics changes from the original
sandbox affect the outlier metrics, it needs to be a second sandbox.

I updated
https://github.com/Recidiviz/recidiviz-data/compare/main...danawillow/id-perf-sandbox
with the changes and the new sandbox command.

Generated sheets with `python -m
recidiviz.tools.insights.performance_review_spreadsheets.generate
--aggregated_metrics_sandbox_prefix dana_id_perf
--outliers_sandbox_prefix dana_id_perf_outliers`

Screenshot of someone with high rates >6 months ago:
https://drive.google.com/file/d/19DxddYIgvjeqrGzn-WvKxZzfmbVHARsl/view?usp=sharing

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#35931
Closes Recidiviz/recidiviz-data#35932

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e2b2b48183bf0facc0d273bbb326b3c5af1402f4