[US_AZ] Exclude NULL officer IDs (Recidiviz/recidiviz-data#36666)
## Description of the change

The full historical version of `AZ_INT_MENTAL_HEALTH_ACTION` has many
`CREATE_USERID` and `UPDT_USERID` values that are the string 'NULL'.
This is causing errors in ingest because that is not the external ID on
an existing `StateStaff` entity. This PR turns those into actual NULLs
before proceeding with entity creation.

Successful sandbox pipeline run:
https://console.cloud.google.com/dataflow/jobs/us-west1/2024-12-23_14_02_05-13713546857166844055;graphView=0?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a0e4b00930f03e1a0a277aeabc06717b8f8d0472