[ID perf] add caseload type for most days in time period (Recidiviz/recidiviz-data#36711)
## Description of the change

Last one! This adds the caseload type that the officer had for the most
number of days in the time period (month or year). As per
https://docs.google.com/spreadsheets/d/1hj8CbLgldF-JuW_y-Ceb0lE4rbSsw-FHlasD2N5-Bo4/edit?disco=AAABZ4fHVS0,
we want the actual caseload type and not the caseload category we show
in Insights. To get the caseload type they had for the highest
proportion of time, I created a new InsightsCaseloadCategoryType in the
sandbox branch, which automatically created aggregated metrics that
represent the proportion of time spent in each caseload type. I had to
change the queries somewhat to allow multiple categories per officer per
category type. Changes there are at
https://github.com/Recidiviz/recidiviz-data/commit/c1f3b49822271d0dd11588106b927a726330d989

`python -m
recidiviz.tools.insights.performance_review_spreadsheets.generate
--aggregated_metrics_sandbox_prefix dana_id_perf
--outliers_sandbox_prefix dana_id_perf_outliers` if you'd like to play
around with the output.

Screenshot of someone with some months with multiple caseload types,
some with 1, and some with none:
https://drive.google.com/file/d/1ewcU9jCnUvvjTPU_ACU8se0jnZ70j2dI/view?usp=sharing

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#35941

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 05af19d07918a1acf17d8cf7756e4c0d3367dea5