expand vitals agg metrics query to cover 6 months of data (Recidiviz/recidiviz-data#36742)
## Description of the change

This is needed for the legacy vitals queries, since that product shows a
graph containing (and allows downloading) 6 months worth of data.

Loaded to sandbox at `dana_vitals_agg_metrics_six_months`. @ageiduschek,
I see that https://github.com/Recidiviz/recidiviz-data/issues/35910 is
still open. When I loaded this it took 38.26 seconds for the most
expensive view. It seems like based on that we can probably merge this
without waiting for that ticket, but wanted to confirm with you first.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#34508

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5a689fb4af71d5063b774a4f43e63609d0b2fb8d