[US_PA][Workflows] Add flags for potential ineligibility to case notes (Recidiviz/recidiviz-data#36196)
## Description of the change

For PA opportunities, there are a few weird gray area cases where we can
kind of check eligibility, but not all the way. For example, someone is
ineligible for admin supervision (per [form
402](https://docs.google.com/document/d/1oyK0ykjfQPFe5ARZGbou6jEzQDFPt-9Y/edit?rtpof=true))
if they have "Former 75 Pa. C.s. 3731 relating to DUI/Controlled
Substance in cases involving bodily injury" on their criminal record. We
can check for statute 3731, but we can't check whether this resulted in
bodily injury. For these sorts of cases, we're going to assume the
client is eligible but flag the charge in the sidebar so agents know to
look into it.

Other cases we want to flag:
1. ADM - 18 Pa. C.S. 5903(4)(5)(6) obscene/sexual material/performance
where the victim is minor (per [form
402](https://docs.google.com/document/d/1oyK0ykjfQPFe5ARZGbou6jEzQDFPt-9Y/edit))-
we already check for 5903(6) in the ineligible offenses criterion
because it always relates to minors, but 5903(4) & (5) could relate to
either adults or minors.
2. SPC - Strangulation cases graded as a felony (per [violent crimes
chart](https://drive.google.com/file/d/1xqkTSMDbj7jU-htkp-1c-qtrYri-dhT8/view?usp=drive_link))
- we can pull the strangulation statute but we don't actually know
whether it was graded as a misdemeanor or felony (per Brenda)
3. ADM/SPC - Some people have special conditions that require an
enhanced supervision for a certain amount of time, so they would not be
eligible for limited supervision during that period. However because
this is a free text field that could depend on a lot of different
factors (e.g. sometimes it's something like: "this client must be on
enhanced supervision for 6 months after release from inpatient care")
I'm not confident we could build this into a criterion accurately.
4. ADM - 35 P.S. 780-113(14)(30) or (37) drug charges should prompt the
user to fill out [form
402a](https://docs.google.com/document/d/13m8k6gNsdjG2M_8S3EG7xZG2Z_UOhbQj/edit?rtpof=true#heading=h.gjdgxs)
and check for sentencing enhancements that could make them ineligible

In order to do this, I create two opportunity-specific case notes
helpers, since the case notes we want to display now vary between
administrative & special circumstances supervision. I also moved the
form 402a helper from the admin opportunity query into
`us_pa_query_fragments`, since that already contained the logic to pull
whether someone has 35 P.S. 780-113(14)(30) or (37) on their record.
Finally, I add special conditions that require enhanced supervision to
the existing special conditions case notes section.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#33420

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 027dd0617c16cc9a69125a27eae39248c65ef359