[US_ND] Fix bug in Docstars staff location periods (Recidiviz/recidiviz-data#36883)
## Description of the change

There are 2 officers who were added to the `docstars_officers` table
with an "Inactive" status before they appeared with an "Active" status.
This is not typically how data is entered into that table, so I was
previously only considering rows where an officer's status changed from
Active to Inactive. This PR tweaks that logic to also consider rows
where an officer goes from Inactive to Active in the raw data.

This will resolve the `current_supervision_staff_missing_district`
validation failure, so I took out the threshold exemption here. I also
removed the exemption for `unidentified_supervision_officer_supervisors`
here since that self-resolved.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0210abfd491041f6168b793082ad4d6679ab5601